+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.6.0 - SoftLock
 - Added SoftLock on PC sheet
+- GmMonitor : Added ability to add or subtract fatigue/strife/void/casualties/panic points on clic.
 
 ## 1.5.0 - FoundryVTT v9 Compatibility
 - Updated the System to FoundryVTT v9.